服务外包比赛/backend/requirements.txt
# ==========================================
# 故障树智能生成系统 - 生产环境依赖
# ==========================================

# FastAPI 核心框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.2.0
python-multipart==0.0.6
aiofiles==23.2.1

# 数据库连接
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
neo4j==5.16.0

# 数据处理与分析
pandas==2.1.4
openpyxl==3.1.2

# 文档处理
python-docx==1.1.0
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.2.0

# AI 模型与知识提取
transformers==4.37.0
torch==2.1.2
accelerate==0.26.0
sentence-transformers==2.3.0
jieba==0.42.1
dashscope==1.14.0

# 图结构与可视化
networkx==3.2.1
matplotlib==3.8.2

# 工具库
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
tenacity==8.2.3
